requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==5.0.6
django-cors-headers==4.4.0
django-rest-auth==0.9.5
django-rest-authtoken==2.1.4
djangorestframework==3.15.1
drf-extra-fields==3.7.0
filetype==1.2.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycparser==2.22
pyOpenSSL==24.2.1
requests==2.32.3
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
zope.interface==6.4.post2
